Mostly some drivetrain work
Code before testing for a P (proportion) controller in the
DriveStraightGyro Command is finished. A matching constant has been
added to the drivetrain class. More could be done to change the gyro
drive straight, e.g. finishing on a bump (reading accelerometer) or by
vision processing (speaking of which, needs to be added and
implemented).