Fleshed out the autonomous. And vision. Vision was added.
This is a "blind" commit; I have no idea whether or not any of the auto
code will work. It is all as placeholders until significant testing can
be completed. Vision processing methods (specifically, UDP classes for
reading the output from a coprocessor) were added with the SocketVision
class. Ladies and gentlemen, we have the ability to SEE!