Added the digital inputs to ports 2 and 3 for the flag limit switches.
Title says it all. Changes made to RoboRio class and RobotMap class.
Limit switches are Digital instead of Analog because they are either
open (low) or closed (high) which can be translated into the boolean
false or true, respectively. Will be used in "flytrap" software to close
the claws automagically.