Made the robot emulate a flytrap
Added two limit switches to the RoboRio (DigitalInputs) and used them in
the default command for the Pneumatics class to schedule the
CloseAllClaws command if either (or both) are triggered. In other words,
if anything hits either limit switch under the claws, they both close.
Again, this is all untested code...