WE DROVE!
Added .gitignore files for all the build jars - nobody wants binary
files, after all... The software is tested and fully functional in
Teleop mode; Autonomous mode still needs to be tested. Buttons may still
change (currently only buttons 1 and 2 or A and B are used).

This is the current software on the robot.