Autonomous Group Worked!
This is the current, stable source for the preseason competition. We (I)
made it, by the skin of our (my) teeth. Has yet to be tested on a bumpy
ground, but by golly it works on carpet. 

Note that there are some inconsistencies that I don't care to redress,
including but not limited to DriveStraightGyro and VisionDriveStraight
requiring oppositely-signed vbus proportions to drive the robot in the
same direction.

I got my mayonnaise, and it tastes like victory.