This is the last commit for this repo
This is the final software we used on our robot in the preseason. If you
want to reuse it, remember to check variable (and object) names, IDs
used in constructors, and to test all your constants IN THE ENVIRONMENT
THEY WILL BE USED IN. Other than that, everything does what its name
implies.